HealthTwinAgent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
groq
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
